[SampleFDO] Refactoring SampleProfileMatcher
